fixed a bunch of exceptions + styling

+++ One-Umbrella.DAL/Repositories/FavoriteRepository.cs
@@ -40,7 +40,7 @@ namespace OneUmbrella.DAL.Repositories
                             RestaurantCity = reader.GetString(reader.GetOrdinal("RESTAURANT_CITY")),
                             RestaurantPostCode = reader.GetString(reader.GetOrdinal("RESTAURANT_POSTCODE")),
                             RestaurantDescription = reader.GetString(reader.GetOrdinal("RESTAURANT_DESCRIPTION")),
-                            RestaurantRating = reader.GetByte(reader.GetOrdinal("RESTAURANT_RATING"))
+                            RestaurantRating = reader.GetDecimal(reader.GetOrdinal("RESTAURANT_RATING"))
                         };
 
                         restaurants.Add(restaurant);
